done with page 2 except carousel
