+++ src/App.js
@@ -1,10 +1,14 @@
 import React from 'react'
-import { Heropage } from 'components/Heropage'
+import { Heropage } from 'components/Heropage/Heropage'
+import { BookClassPage } from 'components/BookClassPage/BookClassPage'
 
 export const App = () => {
   return (
     <div>
-      <Heropage />
+      <div>
+        <Heropage />
+        <BookClassPage />
+      </div>
     </div>
   )
 }
